Rename and move overlay2.tsx
